Replace double with BigDecimal for monetary values and add input validation
- Change all monetary fields from double to BigDecimal for precision
  - BonificoRequestDto.amount: double -> BigDecimal
  - ListaTransactionDto.amount: double -> BigDecimal
  - BalanceDto.balance/availableBalance: double -> BigDecimal
- Add input validation to MessageParserUtil.validateMoneyTransferRequest()
  - Validate amount range (0.01 to 999999999.99)
  - Validate currency format (ISO 4217 - 3 uppercase letters)
  - Validate creditor name (required, max 140 chars)
  - Validate description (required, max 500 chars)
- Extract hardcoded string constants to MessageParserUtil
  - PATTERN_NULLIFEMPTY, PATTERN_NOTRIM, PATTERN_THREE_LETTER_CODE
  - OPERATION_LIS, OPERATION_SAL, OPERATION_BON
  - MIN_TRANSFER_AMOUNT, MAX_TRANSFER_AMOUNT
  - MAX_DESCRIPTION_LENGTH, MAX_CREDITOR_NAME_LENGTH
- Remove jsoniter dependency (use Jackson only)
- Update TransactionValidationUtilTest with 13 new validation tests
- Update DtoSerializationTest to use BigDecimal.compareTo()
- All 59 unit tests pass

+++ src/main/java/it/demo/fabrick/dto/BalanceDto.java
@@ -1,5 +1,6 @@
 package it.demo.fabrick.dto;
 
+import java.math.BigDecimal;
 import java.util.ArrayList;
 
 import lombok.Data;
@@ -10,12 +11,12 @@ public class BalanceDto {
 	public String status;
 	public ArrayList<Error> error;
 	public Payload payload;
-	
+
 	@Data
 	public class Payload{
 	    public String date;
-	    public double balance;
-	    public double availableBalance;
+	    public BigDecimal balance;
+	    public BigDecimal availableBalance;
 	    public String currency;
 	}
